+++ lib-xml/src/main/java/com/orange/ods/xml/component/button/OdsIconToggleButtonsRow.kt
@@ -16,6 +16,7 @@ import android.content.Context
 import android.util.AttributeSet
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.core.content.withStyledAttributes
@@ -27,14 +28,13 @@ import com.orange.ods.xml.utilities.extension.fromXmlAttrValue
 
 class OdsIconToggleButtonsRow @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null) : OdsAbstractComposeView(context, attrs) {
 
-    var icons by mutableStateOf<List<OdsIconToggleButtonsRow.Icon>>(emptyList())
-    var selectedIndex by mutableStateOf(0)
-    var onSelectedIndexChange by mutableStateOf<(Int) -> Unit>({})
+    var iconButtons by mutableStateOf<List<OdsIconToggleButtonsRow.IconButton>>(emptyList())
+    var selectedIconButtonIndex by mutableIntStateOf(0)
     var displaySurface by mutableStateOf(OdsDisplaySurface.Default)
 
     init {
         context.withStyledAttributes(attrs, R.styleable.OdsIconToggleButtonsRow) {
-            selectedIndex = getInt(R.styleable.OdsIconToggleButtonsRow_selectedIndex, 0)
+            selectedIconButtonIndex = getInt(R.styleable.OdsIconToggleButtonsRow_selectedIconButtonIndex, 0)
             displaySurface = OdsDisplaySurface.fromXmlAttrValue(getInteger(R.styleable.OdsIconToggleButtonsRow_displaySurface, 0))
         }
     }
@@ -42,9 +42,8 @@ class OdsIconToggleButtonsRow @JvmOverloads constructor(context: Context, attrs:
     @Composable
     override fun OdsContent() {
         OdsIconToggleButtonsRow(
-            icons = icons,
-            selectedIndex = selectedIndex,
-            onSelectedIndexChange = onSelectedIndexChange,
+            selectedIconButtonIndex = selectedIconButtonIndex,
+            iconButtons = iconButtons,
             displaySurface = displaySurface
         )
     }