+++ lib/src/main/java/com/orange/ods/compose/component/dialog/OdsAlertDialog.kt
@@ -82,7 +82,7 @@ object OdsAlertDialog {
      * @param text Text of the button.
      * @param onClick Will be called when the user clicks the button.
      */
-    class Button(private val text: String, private val onClick: () -> Unit) : OdsComponentContent<Nothing>() {
+    class Button(private val text: String, private val onClick: () -> Unit) : OdsComponentContent<Nothing>(Nothing::class.java) {
 
         @Composable
         override fun Content(modifier: Modifier) {