+++ changelog.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - \[Lib\] Stack `OdsBanner` buttons when texts are too long to fit on one line ([#795](https://github.com/Orange-OpenSource/ods-android/issues/795))
 - \[Lib\] Update `OdsText` APIs and use `OdsText` instead of `Text` whenever possible ([#783](https://github.com/Orange-OpenSource/ods-android/issues/783))
 - \[Lib\] Update `OdsChoiceChipsFlowRow` API ([#794](https://github.com/Orange-OpenSource/ods-android/issues/794))
+- \[Lib\] Update `OdsIconToggleButtonsRow` and `OdsTextToggleButtonsRow` APIs ([#801](https://github.com/Orange-OpenSource/ods-android/issues/801))
 
 ### Fixed